refactor: decompose CampaignChat.test.tsx into focused test suites (#466)
* refactor: decompose CampaignChat.test.tsx into focused test suites

Split the monolithic test file (49 tests) into 8 focused files under
tests/unit/components/CampaignChat/ with a shared helpers.tsx.

- helpers.tsx: CAMPAIGN_ID, sharedTestState, setupFetchMock,
  restoreFetch, openDock, fireMsg, withMembers
- CampaignChat.drawer.test.tsx: TC-01 to TC-13 (13 tests)
- CampaignChat.sse.test.tsx: T2 stream + T11 session events (6 tests)
- CampaignChat.history.test.tsx: T4 history loading (3 tests)
- CampaignChat.unread.test.tsx: T5 unread badge (4 tests)
- CampaignChat.members.test.tsx: T3 member fetching (2 tests)
- CampaignChat.composer.test.tsx: T6/T7/T8 composer (11 tests)
- CampaignChat.visibility.test.tsx: T9 message rendering (4 tests)
- CampaignChat.scene.test.tsx: T10 scene composer (6 tests)

All 49 tests pass unchanged. Uses require('./helpers') inside hoisted
jest.mock callbacks to defer sharedTestState binding past module load.

Closes #452

* chore: update task progress markers

* fix: wrap test suites in describe blocks and fix file:// links in docs

- Wrap all 8 test files in describe('CampaignChat — <concern>') blocks
  per reviewer feedback; jest.mock() calls remain at module scope
- Fix absolute file:/// paths in tasks.md and specs to repo-relative links
- Fix helpers.ts reference to helpers.tsx in tasks.md

* test: address review comments on CampaignChat decomposed tests and restore ProtectedRoute coverage

* refactor: revert unrelated .claude/settings.json changes and clean up unused imports in tests
