+++ .verity/memory/index.md
@@ -4,12 +4,40 @@
 
 > If you are an AI coding agent reading this via CLAUDE.md: scan the catalog below for any node whose title, kind, or file scope is relevant to the task the user just asked you to do. Open the matching files via the Read tool before writing code. Most projects accumulate dozens to hundreds of nodes — do not read them all; pick the few that fit the current change.
 
-## decisions/ (2)
+## decisions/ (16)
 
 - [[n015-model-session-lifecycle-as-its-own-campaign-stream]] — **Model session lifecycle as its own campaign stream event variant**
   *decision* · 84%
 - [[n016-use-sse-campaign-events-for-session-state-updates]] — **Use SSE campaign events for session-state updates instead of polling**
   *decision* · 84%
+- [[n017-invalidate-persisted-dock-height-when-the-viewport]] — **Invalidate persisted dock height when the viewport changes materially**
+  *decision* · 86%
+- [[n018-store-feedback-rate-limits-in-a-ttl-backed-mongodb]] — **Store feedback rate limits in a TTL-backed MongoDB collection**
+  *decision* · 87%
+- [[n019-create-github-issues-server-side-to-keep-the-token]] — **Create GitHub issues server-side to keep the token off the client**
+  *decision* · 91% · scope: `app/api/**/route.ts`
+- [[n020-validate-persisted-chat-sizes-before-reducer-updat]] — **Validate persisted chat sizes before reducer updates**
+  *decision* · 88%
+- [[n021-sanitize-header-derived-text-before-including-it-i]] — **Sanitize header-derived text before including it in GitHub issue bodies**
+  *decision* · 92%
+- [[n022-restrict-feedback-issue-context-urls-to-same-origi]] — **Restrict feedback issue context URLs to same-origin or https**
+  *decision* · 91%
+- [[n023-reject-malformed-feedback-json-before-any-destruct]] — **Reject malformed feedback JSON before any destructuring**
+  *decision* · 86%
+- [[n024-treat-malformed-forwarded-ip-headers-as-non-fatal]] — **Treat malformed forwarded IP headers as non-fatal and fall back to x-real-ip**
+  *decision* · 78%
+- [[n025-run-codacy-finalization-with-if-always]] — **Run Codacy finalization with `if: always()`**
+  *decision* · 84% · scope: `.github/workflows/*.yml`
+- [[n026-use-relative-links-for-spec-references]] — **Use relative links for spec references**
+  *decision* · 82% · scope: `docs/**/*.md`
+- [[n027-keep-repo-workflows-thin-and-centralize-shared-ci]] — **Keep repo workflows thin and centralize shared CI logic in cicd-tooling**
+  *decision* · 82% · scope: `.github/workflows/*.yml`
+- [[n028-use-paragraph-text-not-label-for-non-input-chapter]] — **Use paragraph text, not <label>, for non-input chapter headings**
+  *decision* · 91%
+- [[n029-open-the-chapters-accordion-before-asserting-chapt]] — **Open the chapters accordion before asserting chapter-empty states**
+  *decision* · 84%
+- [[n030-clear-invalid-currentchapterid-before-persisting-c]] — **Clear invalid currentChapterId before persisting campaign edits**
+  *decision* · 90%
 
 ## domain/ (4)
