+++ openspec/changes/decompose-campaign-chat-tests/.openspec.yaml
@@ -1,0 +1,2 @@
+schema: sdd-with-feedback-loop
+created: 2026-07-01
